#63 Edit: removeRepeatList() move to storage.js
