+++ src/store/mutations.js
@@ -1,5 +1,10 @@
 import { getRepeatDates, getRepeatNum } from "@/utils/repeat";
-import { storage } from "@/utils/storage";
+import {
+  storage,
+  createNewKey,
+  createNewGroupId,
+  removeRepeatList,
+} from "@/utils/storage";
 
 const showMemoModal = (state, payload) => {
   state.isDateChanged = false;
@@ -38,6 +43,10 @@ const showMemoModal = (state, payload) => {
 const closeMemoModal = (state) => {
   state.isMemoModalOpen = false;
   state.isEmpty = false;
+};
+
+const checkDateChanged = (state) => {
+  state.isDateChanged = true;
 };
 
 const addOneMemo = (state) => {
@@ -103,23 +112,7 @@ const editRepeatMemo = (state) => {
   let key = createNewKey(state);
   let repeatGroupId = createNewGroupId(state);
 
-  const removedIndex = [];
-  for (let i = 0; i < state.memoList.length; i++) {
-    if (
-      state.currentData.dates <= state.memoList[i].dates &&
-      state.currentData.customData.repeat.groupId ===
-        state.memoList[i].customData.repeat.groupId
-    ) {
-      const indexOfData = state.memoList.findIndex(
-        (item) => item.key === state.memoList[i].key
-      );
-      localStorage.removeItem(state.memoList[i].key);
-      removedIndex.push(indexOfData);
-    }
-  }
-  for (let i = removedIndex.length - 1; i >= 0; i--) {
-    state.memoList.splice(removedIndex[i], 1);
-  }
+  removeRepeatList(state);
 
   const obj = JSON.parse(JSON.stringify(state.currentData));
   obj.customData.repeat.groupId = repeatGroupId;
@@ -149,23 +142,7 @@ const removeOneMemo = (state) => {
 };
 
 const removeRepeatMemo = (state) => {
-  const removedIndex = [];
-  for (let i = 0; i < state.memoList.length; i++) {
-    if (
-      state.currentData.dates <= state.memoList[i].dates &&
-      state.currentData.customData.repeat.groupId ===
-        state.memoList[i].customData.repeat.groupId
-    ) {
-      const indexOfData = state.memoList.findIndex(
-        (item) => item.key === state.memoList[i].key
-      );
-      localStorage.removeItem(state.memoList[i].key);
-      removedIndex.push(indexOfData);
-    }
-  }
-  for (let i = removedIndex.length - 1; i >= 0; i--) {
-    state.memoList.splice(removedIndex[i], 1);
-  }
+  removeRepeatList(state);
 };
 
 const showScheduleList = (state) => {
@@ -200,10 +177,6 @@ const fetchData = (state) => {
   state.memoList = storage.fetch();
 };
 
-const checkDateChanged = (state) => {
-  state.isDateChanged = true;
-};
-
 export {
   fetchData,
   removeOneMemo,
@@ -216,19 +189,4 @@ export {
   editRepeatMemo,
   fetchRepeatCount,
   checkDateChanged,
-};
-
-const createNewGroupId = (state) => {
-  const repeatGroupIdArray = state.memoList.map(
-    (memo) => memo.customData.repeat.groupId
-  );
-  let repeatGroupId =
-    repeatGroupIdArray.length === 0 ? 1 : Math.max(...repeatGroupIdArray) + 1;
-  return repeatGroupId;
-};
-
-const createNewKey = (state) => {
-  const keyArray = state.memoList.map((memo) => memo.key);
-  let key = keyArray.length === 0 ? 1 : Math.max(...keyArray) + 1;
-  return key;
 };
